Cast certain enums to int
This matches the behaviour of MSVC better in cases when there is
a mix of values < 0 and values > 0x7FFFFFFF in the same enum,
which, without this change, prompts GCC to increase enum size to 8 bytes.

Fixes bug #456

(cherry picked from commit 5d61ecd2d087a3a1bdaa9db90270ae0eac121db0)

+++ mingw-w64-headers/include/adoint_backcompat.h
@@ -461,11 +461,11 @@ extern "C" {
   } ConnectModeEnum;
   typedef DECLSPEC_UUID ("00000570-0000-0010-8000-00AA006D2EA4")
   enum RecordCreateOptionsEnum { adCreateCollection = 0x2000,
-    adCreateStructDoc = 0x80000000,
-    adCreateNonCollection = 0,
-    adOpenIfExists = 0x2000000,
-    adCreateOverwrite = 0x4000000,
-    adFailIfNotExists = -1
+    adCreateStructDoc = (int) 0x80000000,
+    adCreateNonCollection = (int) 0,
+    adOpenIfExists = (int) 0x2000000,
+    adCreateOverwrite = (int) 0x4000000,
+    adFailIfNotExists = (int) -1
   } RecordCreateOptionsEnum;
   typedef DECLSPEC_UUID ("00000571-0000-0010-8000-00AA006D2EA4")
   enum RecordOpenOptionsEnum { adOpenRecordUnspecified = -1,
